update route - /cart/items/:productId. add validations

+++ utils/functions.js
@@ -54,6 +54,7 @@ export async function addProduct(res, body) {
   let product = products.find((product) => product.id === productId);
   if (!isfound(res, customer)) return;
   if (!isfound(res, product)) return;
+  if (!isExists(res, product)) return;
   if (!isStuckFull(res, product.stock)) return;
   customer.cart.push({ productId, quantuty });
   await writeFile("./database/customers.json", customers);
@@ -61,17 +62,27 @@ export async function addProduct(res, body) {
   return "product added";
 }
 
-export async function deleteProduct(productId, customerId) {
+export async function deleteProduct(res, productId, customerId) {
+  if (!isExists(res, customerId)) return;
+  if (!isExists(res, productId)) return;
+  if (!isNumber(res, +customerId)) return;
+  if (!isNumber(res, +productId)) return;
   const customers = await readFile("./database/customers.json");
   const products = await readFile("./database/products.json");
   let customer = customers.find(
     (customer) => customer.customerId === customerId,
   );
-  let product = products.find((product) => product.id === productId);
-  product.stock += 1;
+  let productFordeleting = customer.cart.find(
+    (product) => +productId === +productId,
+  );
+  if (!isfound(res, customer)) return;
+  if (!isfound(res, productFordeleting)) return;
+  if (!isExists(res, productFordeleting)) return;
+  customer.cart.pop(productFordeleting);
+  // product.stock += 1;
   await writeFile("./database/customers.json", customers);
   await writeFile("./database/products.json", products);
-  return;
+  return "product deleted";
 }
 
 export async function getBalance(customerId) {
